% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stutils.R
\name{dist.pw}
\alias{dist.pw}
\title{Pair-wise distance calculation}
\usage{
dist.pw(df)
}
\arguments{
\item{df}{Input data frame with (x, y) as spatial coordiantes
and z as an variable}
}
\value{
The column \code{r} represents Euclidean distance;
    columns \code{z1} and \code{z2} preserve corresponding \code{z}
    values of two points.
}
\description{
Calculate pair-wise distance metrics of row-wise locations.
}
\examples{
dist.pw(data.frame(x=seq(6), y=seq(6), z=runif(6)))
}

