% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd.hf}
\alias{gcd.hf}
\title{Great circle distance calculation using the Haversine formula}
\usage{
gcd.hf(long1, lat1, long2, lat2)
}
\arguments{
\item{long1,lat1}{Longitude and latitude coordinates of point 1.}

\item{long2,lat2}{Lontitude and latitude coordinates of point 2.}
}
\description{
Calculates the geodesic distance between two points specified by radian
latitude/longitude using the Haversine formula (hf).
}
\examples{
gcd.hf(120.0, 30.0, 120.5, 30.5)
}
\seealso{
Other Great Circle Distance Calculation: \code{\link{gcd.slc}};
  \code{\link{gcd.vif}}
}

