% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd.vif}
\alias{gcd.vif}
\title{Great circle distance calculation using Vincenty inverse formula
for ellipsoids}
\usage{
gcd.vif(long1, lat1, long2, lat2)
}
\arguments{
\item{long1,lat1}{Longitude and latitude coordinates of point 1.}

\item{long2,lat2}{Lontitude and latitude coordinates of point 2.}
}
\description{
Calculates the geodesic distance between two points specified by radian
latitude/longitude using Vincenty inverse formula for ellipsoids (vif)
}
\examples{
gcd.vif(120.0, 30.0, 120.5, 30.5)
}
\seealso{
Other Great Circle Distance Calculation: \code{\link{gcd.hf}};
  \code{\link{gcd.slc}}
}

