% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/omniR.R
\docType{package}
\name{omniR}
\alias{omniR}
\alias{omniR-package}
\title{omniR: useful functions for spatio-temporal data analysis}
\description{
Package containing useful functions in my researches of human
behavior science, mostly for spatio-temporal data
}

