% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{rangebinc}
\alias{rangebinc}
\title{Bined range}
\usage{
rangebinc(x, n)
}
\arguments{
\item{x}{A vector to bin}

\item{n}{Number of bins}
}
\description{
Bin the range of given vector into n itervals, and
return the center of each interval.
}
\examples{
rangebinc(seq(1:10), 3)
}
\seealso{
Other converting functions: \code{\link{vbingrid}};
  \code{\link{vbin}}
}

