% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stutils.R
\name{stscale}
\alias{stscale}
\title{Spatio-temporal scaling}
\usage{
stscale(scoord, tcoord, value, alpha = 0.5)
}
\arguments{
\item{scoord}{A vector defining spatial dimension of random field.}

\item{tcoord}{A vector defining temporal dimension of random field.}

\item{value}{A vector giving values in coorsponding random field slot.}

\item{alpha}{A tuning parameter controling the weight of spatial or temporal
coordinate in the scaled value.}
}
\description{
Scale the value along two coordinates simultaneously, e.g.,
along space and time in a random field models.
}
\examples{
scoord <- rep(seq(6), 2)
tcoord <- rep(c(1,2), each=6)
value <- runif(6 * 2)
stscale(scoord, tcoord, value)
}

