% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{vbin}
\alias{vbin}
\title{Bined vector}
\usage{
vbin(x, n, type = "i")
}
\arguments{
\item{x}{A vector to bin}

\item{n}{Number of bins}

\item{type}{Interval representation form, c("i", "c")}
}
\value{
Interval index if type is "i"; interval center point if
    type is "c".
}
\description{
Bin a vector into n intervals in regard with its value range.
The vector x is split into n bins within [min(x), max(x)],
and bin index is given by checking the bin [bin_min, bin_max)
into which data points in x fall.
}
\examples{
vbin(seq(1:10), 3)
}
\seealso{
Other converting functions: \code{\link{rangebinc}};
  \code{\link{vbingrid}}
}

