% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert.R
\name{vbingrid}
\alias{vbingrid}
\title{Bined matrix of Random Field}
\usage{
vbingrid(x, y, v, FUN = mean, nx = 50, ny = 50, na = NA)
}
\arguments{
\item{x,y,v}{Random fields of (x, y) takeing value z. They must be
vectors of the same length.}

\item{FUN}{A fun to calculate aggretated value in each matrix bin.}

\item{nx,ny}{Number of bins in each dimension.}

\item{na}{Replacemnet for NA value in matrix bins.}
}
\description{
Generate bined matrix given vectors of a random field.
}
\examples{
vbingrid(seq(1:20), seq(21:30), runif(20), na=0)
}
\seealso{
Other converting functions: \code{\link{rangebinc}};
  \code{\link{vbin}}
}

