% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/omniR.R
\docType{package}
\name{omniR}
\alias{omniR}
\alias{omniR-package}
\title{omniR (R data analysis at OMNILab)}
\description{
Package containing useful functions to ease my researches of human
behavior science and spatio-temporal data analysis.
}

